replaced BottomNav with custom NewNav
