+++ navigation/NewNavigator.js
