+++ screens/FormScreen.js
@@ -6,16 +6,16 @@ import { RectButton, ScrollView } from 'react-native-gesture-handler';
 
 import styles from '../allStyles';
 
-var DomParser = require('react-native-html-parser').DOMParser
 
 
 let xhtmlTestOutput = "-xhtml-";
 
 
-export default function FormScreen() {
+export default function FormScreen({route,navigation}) {
 
-  requestxhtml();
+  //xhtmlTestOutput = route.params.xhtmltest;
 
+ // requestxhtml();
   return (
     <ScrollView style={styles.form_container} contentContainerStyle={styles.form_contentContainer}>
       <OptionButton
@@ -48,44 +48,60 @@ export default function FormScreen() {
   );
 }
 
+// //#region REACT FETCH functions      **call fetchFromStimulus();
+// function fetchFromStimulus() {
+//   var t = async () => {
+//     var url = 'https://www.glenwoodoilandgasinc.com/appinfo';
+//     var _result = (await fetch(url)).resp.text();//'no result get rekt';
+//   // await fetch(url).then((resp)=>{ return resp.text() }).then((text)=>{ console.log(text) })
+//     console.log("fetch result be like5: " + _result);
+//   }
+//   t();
+// }
+// //#endregion REACT FETCH functions
 
-function justPrintResponseText(){
-  console.log(this.responseText);
-}
+// //#region [ rgba (10, 100, 10, 0.2) ] XML HTTP REQUEST functions  ** to use, call requestxhtml();
 
-/**
- * takes string and looks for > then looks for < and substrings in between them
- * @param {string only} _string 
- */
-function takeOutMLtags(_string){
-  if(_string.includes(">"))
-  {
-    var returnString = _string.substring(_string.indexOf('>')+1, _string.lastIndexOf('<'));
-    console.log("TAKE OUT RESULT:" + returnString);
-    return returnString;
-  }
-  else return "INVALID_STRING - functionn:takeOutMLtags";
-}
-
-function parseResponseText(){
-  let _response = this.responseText;
-  let xhtmlDom = new DomParser().parseFromString(_response,'text/html');
-  var before = "BEFORE:" + xhtmlDom.querySelect('#home #paragraph');
-  console.log(before);
-  takeOutMLtags(before);
-}
+// var DomParser = require('react-native-html-parser').DOMParser
 
 
+// function justPrintResponseText(){
+//   console.log(this.responseText);
+// }
 
-function requestxhtml() {
-  var xmlRequest = new XMLHttpRequest();
-  xmlRequest.addEventListener("load", parseResponseText);
-  xmlRequest.open("GET","https://www.glenwoodoilandgasinc.com/appinfo");
-  xmlRequest.send();
-  console.log("Success TOM");
-  //return xmlReresponseText;
-}
+// /**
+//  * takes string and looks for > then looks for < and substrings in between them
+//  * @param {string} _string must have exactly one <tag>content</tag>
+//  */
+// function takeOutMLtags(_string){
+//   if(_string.includes(">"))
+//   {
+//     var returnString = _string.substring(_string.indexOf('>')+1, _string.lastIndexOf('<'));
+//     console.log("TAKE OUT RESULT:" + returnString);
+//     return returnString;
+//   }
+//   else return "INVALID_STRING - function:takeOutMLtags";
+// }
 
+// function parseResponseText(){
+//   let _response = this.responseText;
+//   let xhtmlDom = new DomParser().parseFromString(_response,'text/html');
+//   var before = "BEFORE:" + xhtmlDom.querySelect('#home #paragraph');
+//   console.log(before);
+//   this.state = {updated: true};
+// }
+
+
+
+// function requestxhtml() {
+//   var xmlRequest = new XMLHttpRequest();
+//   xmlRequest.addEventListener("load", parseResponseText);
+//   xmlRequest.open("GET","https://www.glenwoodoilandgasinc.com/appinfo");
+//   xmlRequest.send();
+//   console.log("Success TOM");
+//   //return xmlReresponseText;
+// }
+// //#endregion XML HTTP REQUEST functions
 
 
 function OptionButton({ icon, label, onPress, isLastOption }) {
